MultiCurrencyMoney/Franc.cs
using System;
using System.Collections.Generic;
using System.Text;

namespace MultiCurrencyMoney
{
    public class Franc
    {
        private int _amount;
        public Franc(int amount)
        {
            _amount = amount;
        }

        public Franc Times(int multiplier)
        {
            return new Franc(_amount * multiplier);
        }

        public override bool Equals(object obj)
        {
            if (obj == null)
                return false;

            if (this.GetType() != obj.GetType())
                return false;

            Franc Franc = (Franc)obj;
            if (this._amount == Franc._amount)
                return true;
            else
                return false;
        }
    }
}
